market_ai_kit/requirements.txt
yfinance
pandas
numpy
scipy
ta
requests
pyyaml
